STRWEB/LR3/requirements.txt
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
fonttools==4.58.5
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
